tree,catalog/colinfo: move some code from tree to colinfo
This is a small and isolated code movement change as I work to untangle
resolution from the the tree package.

Release note: None
