+++ pkg/sql/opt/optbuilder/BUILD.bazel
@@ -93,6 +93,8 @@ go_test(
     embed = [":optbuilder"],
     deps = [
         "//pkg/settings/cluster",
+        "//pkg/sql/catalog/colinfo",
+        "//pkg/sql/catalog/colinfo/colinfotestutils",
         "//pkg/sql/opt/cat",
         "//pkg/sql/opt/memo",
         "//pkg/sql/opt/testutils",
@@ -103,7 +105,6 @@ go_test(
         "//pkg/sql/sem/builtins",
         "//pkg/sql/sem/tree",
         "//pkg/sql/types",
-        "//pkg/testutils/sqlutils",
         "//pkg/util/leaktest",
         "@com_github_cockroachdb_datadriven//:datadriven",
     ],
